+++ server/src/main/java/org/opensearch/node/NodeService.java
@@ -83,6 +83,7 @@ public class NodeService implements Closeable {
     private final IndexingPressureService indexingPressureService;
     private final AggregationUsageService aggregationUsageService;
     private final SearchBackpressureService searchBackpressureService;
+    private final ClusterService clusterService;
 
     private final Discovery discovery;
 
@@ -123,6 +124,7 @@ public class NodeService implements Closeable {
         this.indexingPressureService = indexingPressureService;
         this.aggregationUsageService = aggregationUsageService;
         this.searchBackpressureService = searchBackpressureService;
+        this.clusterService = clusterService;
         clusterService.addStateApplier(ingestService);
     }
 
@@ -174,7 +176,8 @@ public class NodeService implements Closeable {
         boolean scriptCache,
         boolean indexingPressure,
         boolean shardIndexingPressure,
-        boolean searchBackpressure
+        boolean searchBackpressure,
+        boolean clusterManagerThrottling
     ) {
         // for indices stats we want to include previous allocated shards stats as well (it will
         // only be applied to the sensible ones to use, like refresh/merge/flush/indexing stats)
@@ -197,7 +200,8 @@ public class NodeService implements Closeable {
             scriptCache ? scriptService.cacheStats() : null,
             indexingPressure ? this.indexingPressureService.nodeStats() : null,
             shardIndexingPressure ? this.indexingPressureService.shardStats(indices) : null,
-            searchBackpressure ? this.searchBackpressureService.nodeStats() : null
+            searchBackpressure ? this.searchBackpressureService.nodeStats() : null,
+            clusterManagerThrottling ? this.clusterService.getClusterManagerService().getThrottlingStats() : null
         );
     }